Response yapıları düzenlendi.

+++ src/main/java/com/gokhan/bitcode/service/UserService.java
@@ -1,11 +1,11 @@
 package com.gokhan.bitcode.service;
 
+import com.gokhan.bitcode.ApiResponse;
 import com.gokhan.bitcode.entity.UserEntity;
 import com.gokhan.bitcode.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class UserService {
@@ -16,16 +16,26 @@ public class UserService {
         this.userRepository = userRepository;
     }
 
-    public List<UserEntity> getAllUsers() {
-        return userRepository.findAll();
+    public ApiResponse<List<UserEntity>> getAllUsers() {
+        return ApiResponse.success(userRepository.findAll());
     }
 
-    public Optional<UserEntity> getUserById(Long id) {
-        return userRepository.findById(id);
+    public ApiResponse<UserEntity> getUserById(Long id) {
+        return userRepository.findById(id)
+                .map(ApiResponse::success)
+                .orElse(ApiResponse.userNotFound());
     }
 
-    public void deleteUserById(Long id) {
-        userRepository.deleteById(id);
+    public ApiResponse<Void> deleteUserById(Long id) {
+        if (!userRepository.existsById(id)) {
+            return ApiResponse.userNotFound();
+        }
+        try {
+            userRepository.deleteById(id);
+            return ApiResponse.success(null, 204);
+        } catch (Exception e) {
+            return ApiResponse.badRequest("BIT-1003", "Kullanıcı silinirken bir hata oluştu.");
+        }
     }
 }
